metric error

+++ generateTraffic.sh
@@ -11,7 +11,7 @@ host=$1
 # Trap SIGINT (Ctrl+C) to execute the cleanup function
 cleanup() {
   echo "Terminating background processes..."
-  kill $pid1 $pid2 $pid3 $pid4 $pid5
+  kill $pid1 $pid2 $pid3 $pid4 $pid5 $pid6
   exit 0
 }
 trap cleanup SIGINT
@@ -19,6 +19,11 @@ trap cleanup SIGINT
 # Wrap curl command to return HTTP response codes
 execute_curl() {
   curl -s -o /dev/null -w "%{http_code}" "$@"
+}
+
+# Function to register a user
+register() {
+  curl -s -o /dev/null -X POST "$host/api/auth" -d "{\"name\":\"$1\", \"email\":\"$2\", \"password\":\"$3\"}" -H 'Content-Type: application/json'
 }
 
 # Function to login and get a token
@@ -31,6 +36,11 @@ login() {
   fi
   echo "$response" | grep -o '"token":"[^"]*"' | cut -d'"' -f4
 }
+
+# --- INITIAL REGISTRATION ---
+echo "Registering default test users..."
+register "Franchise Owner" "f@jwt.com" "franchisee"
+register "Hungry Diner" "d@jwt.com" "diner"
 
 # Simulate a user requesting the menu every 3 seconds
 while true; do
@@ -53,8 +63,10 @@ while true; do
   token=$(login "f@jwt.com" "franchisee")
   echo "Login franchisee..." $( [ -z "$token" ] && echo "false" || echo "true" )
   sleep 110
-  result=$(execute_curl -X DELETE "$host/api/auth" -H "Authorization: Bearer $token")
-  echo "Logging out franchisee..." $result
+  if [ -n "$token" ]; then
+    result=$(execute_curl -X DELETE "$host/api/auth" -H "Authorization: Bearer $token")
+    echo "Logging out franchisee..." $result
+  fi
   sleep 10
 done &
 pid3=$!
@@ -62,12 +74,12 @@ pid3=$!
 # Simulate a diner ordering a pizza every 50 seconds
 while true; do
   token=$(login "d@jwt.com" "diner")
-if [ -z "$token" ]; then
-  echo "Login failed, skipping order..."
-  sleep 10
-  continue
-fi
-  echo "Login diner..." $( [ -z "$token" ] && echo "false" || echo "true" )
+  if [ -z "$token" ]; then
+    echo "Login failed, skipping order..."
+    sleep 10
+    continue
+  fi
+  echo "Login diner... true"
   result=$(execute_curl -X POST "$host/api/order" -H 'Content-Type: application/json' -d '{"franchiseId": 1, "storeId":1, "items":[{ "menuId": 1, "description": "Veggie", "price": 0.05 }]}' -H "Authorization: Bearer $token")
   echo "Bought a pizza..." $result
   sleep 20
@@ -80,7 +92,12 @@ pid4=$!
 # Simulate a failed pizza order every 5 minutes
 while true; do
   token=$(login "d@jwt.com" "diner")
-  echo "Login hungry diner..." $( [ -z "$token" ] && echo "false" || echo "true" )
+  if [ -z "$token" ]; then
+    echo "Login hungry diner failed, skipping..."
+    sleep 10
+    continue
+  fi
+  echo "Login hungry diner... true"
 
   items='{ "menuId": 1, "description": "Veggie", "price": 0.05 }'
   for (( i=0; i < 21; i++ ))
@@ -96,6 +113,14 @@ while true; do
 done &
 pid5=$!
 
+# Simulate a new user registering every 60 seconds
+while true; do
+  random_string=$(tr -dc A-Za-z0-9 </dev/urandom | head -c 8)
+  result=$(execute_curl -X POST "$host/api/auth" -H 'Content-Type: application/json' -d "{\"name\":\"test_$random_string\", \"email\":\"test_$random_string@jwt.com\", \"password\":\"diner\"}")
+  echo "Registered new user test_$random_string..." $result
+  sleep 60
+done &
+pid6=$!
 
 # Wait for the background processes to complete
-wait $pid1 $pid2 $pid3 $pid4 $pid5
+wait $pid1 $pid2 $pid3 $pid4 $pid5 $pid6
